Break down handler.rs into separate files
